First part of old checkout cleanup
